разбил по действиям
